Update v1.3: Better server connections, added server reconnects, and client disconnect handling
Update v1.3: Better server connections, added server reconnects, and client disconnect handling

Client: On start, the client no longer crashes if it is unable to connect to the server
Client: No longer crashes on losing connection to server
Client: Now will periodically attempt to connect to the server (every 1 second) if connection isn't established or has been lost.
Client: Added ID based on system
Client: Added new system variables that can be passed to the server (OS, Device Name, Version, Architecture)
Client: On start, it sends initilization data (System info, resources, & screenshot) to the server to hold and send to all web clients who happen to be already trying to view this device (saved tab, or in the future saved devices in the down state)

Server: Allows for client reconnects and updates variables (Resources and Screenshot)
Server: Fixed issue when python client disconnects and reconnects
Server: On web client disconnect, it now detects the closed socket state and removes it. And adds it back on reconnecting
Server: Fixed forced shutdown on sending a message to a closed socket. This is a last resort and should not be needed, and will throw an error if it reaches it, without closing the server.

Web Client: Now can handle server not being online initially, and auto reconnects to server if it ever loses connection
Web Client: Web client now automatically gets cached data upon viewing remote client.

Future Ideas:

Store connected web clients in a set on the python client, and have the server send updates for new clients and disconnects to the python client, so we don't have to send through the server
So the python client instead of going like this Web Client > Server > Python Client > Server > Web Client, we can do Web Client > Server > Python Client > Web client.
Additionally, to take it one step further, is we could have the web client store its connected python client, although this isn't the best idea to let it interface directly, but we could then go like Web Client > Python Client > Web Client, removing 2 data transfers.
Of course, these will then add more layers of complexity and less security for future versions. I think it's best allow only communication to the server, so the server can process the data and we can make the web clients and python clients not publicly accessable. Plus, we can store more information on the server like cached data for new clients connecting without making more overhead for the python clients

Currently there is no way to store data if the server is down. So when the server is started up, if the client is down, you can't see its prior information. This is not seen as critical at the moment, but rather more features will be my priority for the time being.