Update v1.5 - File Viewer
File Viewer:

Added functionality for the file viewer across the clients and server. Upon opening the web client you'll get a list of drives on the machine, then you can select which one you'd like to view, and see all files in the directory. You can open folders to view its items as well.
Back button allows you to go up a directory up to the root. This is handled on the python client side to determine the parent directory or itll return all the drives available if you back out of any drive
Cannot open or download files yet

First load information sends over the name of drives, along with current and max storage for each
Folder information sends current directory and for every file or folder, the name, and file size (or says "Folder")

Fixed:

Server: Fixed restarts of clients creating duplicates because of programming error regarding checking for existing clients
Server: Fixed error with sending screenshots over json on updated data transmit methods

Python Client: Improved if statements for recieving commands from previous suboptimal code

Future Updates: 

Currently I have disabled sending over sizes of folders as it could take a while to calculate each one, but the code exists to do it. In the future we could add a toggle to send that along with it.
File sorting